+++ backend/src/test/java/kernel360/techpick/folder/service/parser/ParserTestCase.java
@@ -8,14 +8,12 @@ public class ParserTestCase {
 	public static final String CASE_INVALID_B = "[{]";
 	// 빈 정보일 경우
 	public static final String CASE_INVALID_C = "[{}]";
-	// 필요한 정보가 타입이 잘못된 경우 (id != String)
-	public static final String CASE_INVALID_D = "[{id:\"STR\", type:folder, folderId:1, name:hi, children:null}]";
 	// 필요한 정보가 타입이 잘못된 경우 (TYPE = FOLDER | PICK)
-	public static final String CASE_INVALID_E = "{id:\"STR\", type:XXXX, folderId:1, name:\"폴더\", children: null}";
+	public static final String CASE_INVALID_D = "{id:\"STR\", type:XXXX, folderId:1, name:\"폴더\", children: null}";
 	// 필요한 정보가 타입이 잘못된 경우 (children != array)
-	public static final String CASE_INVALID_F = "{id:\"STR\", type:folder, folderId:1, name:\"폴더\", children: {foo:bar}}";
+	public static final String CASE_INVALID_E = "{id:\"STR\", type:folder, folderId:1, name:\"폴더\", children: {foo:bar}}";
 	// id 값이 문자 열이 아닌 경우 (프론트엔드가 string으로 처리)
-	public static final String CASE_INVALID_K = "{id:3, type:XXXX, folderId:1, name:\"폴더\", children: {foo:bar}}";
+	public static final String CASE_INVALID_F = "{id:3, type:XXXX, folderId:1, name:\"폴더\", children: {foo:bar}}";
 	// 필요한 정보가 일부 없는 경우 (type이 없음)
 	public static final String CASE_INVALID_G = "{id:\"STR\", folderId:1, children: null}";
 	// 빈 배열은 OK
